requirements.txt
google-cloud-firestore==1.4.0
google-cloud-bigquery==1.17.0
pandas==0.25.0
pyarrow==0.14.1
